requirements.txt
gradio>=5.0
httpx>=0.27
uvicorn>=0.30
sqlite-vec>=0.1.3
langfuse>=3.0
prometheus-client>=0.20
PyMuPDF>=1.24
pyyaml>=6.0
websockets>=12.0

# LangGraph agent backend (used when AGENT_BACKEND=langgraph)
langchain>=1.2.3
langgraph>=1.1.0
langchain-openai>=0.3.0
